fix(storage): harden bounded bulk reingest
